+++ src/.internal/charts/xy/XYChartDefaultTheme.ts
@@ -181,7 +181,8 @@ export class XYChartDefaultTheme extends Theme {
 		r("XYCursor").setAll({
 			behavior: "none",
 			layer: 20,
-			exportable: false
+			exportable: false,
+			snapToSeriesBy: "xy"
 		});
 
 		{